requirements.txt
torch
transformers
faiss-cpu
numpy
tqdm
matplotlib
